requirements_approx.txt
evdev

pi-plates
spidev
RPi.GPIO (or python3-rpi.gpio in apt)
pyyaml
